- Implémenté ValidationContext.Refresh et UpdateValidity. - Ajouté CommandState.SplitGroupNames et JoinGroupNames. - Ajouté ValidationGroups à Visual, ainsi que des propriétés relatives à la validation, anciennement gérées par Widget. - L'identificateur de Visual est maintenant unique, vraiment (long à la place de short).
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2005-cmd-may-2006@5372 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Common.Widgets/Layouts/LayoutContext.cs
@@ -503,7 +503,7 @@ namespace Epsitec.Common.Widgets.Layouts
 			
 			public override int GetHashCode()
 			{
-				return this.visual.VisualSerialId;
+				return (int) this.visual.VisualSerialId;
 			}
 			public override bool Equals(object obj)
 			{
